fix: Align disk-size popover menu columns like the wizard
## Summary
- The "Create New Disk" / "Create New Removable Disk" popover's size menu used
  plain titles, so the number/unit columns didn't line up the way the wizard's
  disk popup does. Use the same tab-stopped attributed titles in both.

## Changes
- Add shared `diskSizeMenuItemTitle(_:)` (Views/Shared/DiskSizeMenuTitle.swift):
  number right-aligned + unit left-aligned to shared tab stops
- DiskSizePopoverContentViewController: set each size item's attributedTitle
- ResourceConfigContentViewController: use the shared helper; remove its private
  `diskItemTitle(_:)` copy

## Test plan
- [x] make build
- [x] make lint --strict clean
- [ ] Manual: the Create-disk popover's size menu lines up number/unit columns
  identically to the wizard's Disk Size popup
